+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import NotFound from './components/pages/NotFound';
-import Dragons from './components/dragons';
+import Dragons from './components/dragon/dragons';
 import MissionsContainer from './components/pages/MissionsContainer';
 import Rockets from './components/rockets/Rockets';
 import NavBar from './components/NavBar/NavBar';
