+++ src/components/dragon/DragonElement.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
-import { reserveDragon, unreserveDragon } from '../redux/dragons/Dragons';
+import { reserveDragon, unreserveDragon } from '../../redux/dragons/Dragons';
 
 const DragonElement = (props) => {
   const dispatch = useDispatch();
