+++ src/components/dragon/dragons.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { getDragons } from '../redux/dragons/Dragons';
+import { getDragons } from '../../redux/dragons/Dragons';
 import DragonElement from './DragonElement';
 import './dragons.css';
